+++ app/public/index.php
@@ -4,41 +4,58 @@ require_once __DIR__ . '/../src/helpers/url_helper.php';
 
 $routeConfig = require __DIR__ . '/../config/routes.php';
 
-function loadView(string $viewPath): void
+function loadController(string $controllerName, string $method): void
 {
-    global $routeConfig;
-    $fullPath = $routeConfig["view_dir"] . '/' . $viewPath;
-    
-    if (file_exists($fullPath)) {
-        include($fullPath);
-        return;
+    $controllerClass = $controllerName;
+
+    if (!class_exists($controllerClass)) {
+        throw new RuntimeException("Controlador no encontrado: $controllerClass");
     }
-    
-    throw new RuntimeException("View not found: $fullPath");
+
+    $controller = new $controllerClass();
+
+    if (!method_exists($controller, $method)) {
+        throw new RuntimeException("Método no encontrado: $controllerClass::$method");
+    }
+
+    $controller->$method();
 }
+
+// Iniciar sesión al principio
+session_start();
 
 // Obtener la ruta solicitada
 $requestUri = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
 $pathInfo = rtrim($requestUri, '/') ?: '/';
 
 try {
-    // Buscar la ruta que coincida con el path solicitado
     $foundRoute = false;
+
     foreach ($routeConfig['routes'] as $route) {
         $routePath = rtrim($route['path'], '/') ?: '/';
-        
+
         if ($routePath === $pathInfo) {
-            loadView($route['view']);
+            if (isset($route['controller']) && isset($route['method'])) {
+                // Nueva arquitectura con controladores
+                loadController($route['controller'], $route['method']);
+            } else {
+                // Arquitectura antigua (para migración gradual)
+                loadView($route['view']);
+            }
             $foundRoute = true;
             break;
         }
     }
-    
+
     if (!$foundRoute) {
         http_response_code(404);
-        loadView($routeConfig['routes']['404']['view']);
+        include __DIR__ . '/../src/view/errors/404.php';
     }
+
 } catch (RuntimeException $e) {
     http_response_code(500);
     echo "Error: " . $e->getMessage();
+} catch (Exception $e) {
+    http_response_code(403);
+    include __DIR__ . '/../src/view/errors/403.php';
 }
